requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.1
chromadb>=0.4.15
sentence-transformers>=2.2.2
selenium>=4.15.2
pymupdf>=1.23.8
python-multipart>=0.0.6
pydantic>=2.5.0
numpy>=1.24.3
pandas>=2.1.3
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
setuptools>=65.0.0
wheel>=0.38.0
google-generativeai>=0.5.2